+++ app/dict_utils/__init__.py
@@ -1,0 +1,6 @@
+def get_or_create(aDict, key, default=lambda: []):
+    arr = aDict.get(key, default())
+    aDict[key] = arr
+    return arr
+
+
